Refactor: change to fit coding standards
